feat(move_piece): moved move_piece into a one
