+++ andy_api/api/intent_processing/utils.py
@@ -15,7 +15,7 @@ INTENT_NAME_BASE = "projects/chess-master-andy-mhyo/agent/intents/"
 # The name of each type is separated by a space
 RESPONSE_TYPES = Enum(
     "RESPONSE_TYPES",
-    "HELLO FALLBACK CHOOSE_SIDE MOVE_PIECE_FROM MOVE_PIECE_TO WAKE_UP_PROMPT WAKE_UP_FOLLOW_UP_YES WAKE_UP_FOLLOW_UP_NO ERROR"
+    "HELLO FALLBACK CHOOSE_SIDE MOVE_PIECE WAKE_UP_PROMPT WAKE_UP_FOLLOW_UP_YES WAKE_UP_FOLLOW_UP_NO ERROR"
 )
 
 
@@ -23,8 +23,7 @@ INTENT_MAPPING = {
     INTENT_NAME_BASE + "2ab3d889-b6eb-494e-b822-9992da79280c": RESPONSE_TYPES.FALLBACK,
     INTENT_NAME_BASE + "fc298129-5845-44dc-a976-b7d6ca2f14c3": RESPONSE_TYPES.HELLO,
     INTENT_NAME_BASE + "6fafe557-d27b-41e7-bef0-204a87036e2c": RESPONSE_TYPES.CHOOSE_SIDE,
-    INTENT_NAME_BASE + "3d732581-d539-44cf-8fce-3a9b18dd7e59": RESPONSE_TYPES.MOVE_PIECE_FROM,
-    INTENT_NAME_BASE + "a252d16d-dcdb-47af-a049-7b5bfcdd2618": RESPONSE_TYPES.MOVE_PIECE_TO,
+    INTENT_NAME_BASE + "67bf1b70-c4f3-44e5-976e-960837acff06": RESPONSE_TYPES.MOVE_PIECE,
     INTENT_NAME_BASE + "2b614d03-2366-4878-b22d-86df4003138d":
     RESPONSE_TYPES.WAKE_UP_PROMPT,
     INTENT_NAME_BASE + "f43bfd0d-940f-42b4-bee7-87e95334b0ae":
